requirements.txt
accelerate==0.26.1
aiohttp==3.9.0
aiosignal==1.3.1
annotated-types==0.6.0
ansible==6.7.0
ansible-core==2.13.13
ansible-vault==2.1.0
antlr4-python3-runtime==4.9.3
anyio==4.2.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
async-timeout==4.0.3
attrs==23.2.0
Babel==2.13.1
backoff==2.2.1
base58==2.1.1
beautifulsoup4==4.12.2
bittensor==6.6.1
black==23.7.0
bleach==6.1.0
certifi==2023.11.17
cffi==1.16.0
cfgv==3.4.0
charset-normalizer==3.3.2
click==8.1.7
comm==0.2.0
contourpy==1.2.0
cryptography==41.0.3
cycler==0.12.1
cytoolz==0.12.2
datasets==2.16.1
ddt==1.6.0
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
diffusers==0.25.1
dill==0.3.7
diskcache==5.6.3
distlib==0.3.8
distro==1.7.0
ecdsa==0.18.0
entrypoints==0.4
eth-hash==0.5.2
eth-keys==0.4.0
eth-typing==3.5.2
eth-utils==2.3.1
exceptiongroup==1.1.3
executing==2.0.1
fastapi==0.99.1
fastjsonschema==2.18.1
filelock==3.13.1
fonttools==4.47.2
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2023.10.0
ftfy==6.1.3
fuzzywuzzy==0.18.0
gunicorn==21.2.0
h11==0.14.0
httpcore==1.0.2
httplib2==0.20.2
httpx==0.26.0
huggingface-hub==0.20.2
identify==2.5.33
idna==3.6
imagehash==4.3.1
imageio==2.33.1
importlib-metadata==4.6.4
iniconfig==2.0.0
ipykernel==6.26.0
ipython==8.17.2
ipython-genutils==0.2.0
ipywidgets==8.1.1
isoduration==20.11.0
jedi==0.19.1
jeepney==0.7.1
Jinja2==3.1.2
json5==0.9.14
jsonpointer==2.4
jsonschema==4.19.2
jsonschema-specifications==2023.7.1
jupyter-archive==3.4.0
jupyter-contrib-core==0.4.2
jupyter-contrib-nbextensions==0.7.0
jupyter-events==0.9.0
jupyter-highlight-selected-word==0.2.0
jupyter-lsp==2.2.0
jupyter-nbextensions-configurator==0.6.3
jupyter_client==7.4.9
jupyter_core==5.5.0
jupyter_server==2.10.0
jupyter_server_terminals==0.4.4
jupyterlab==4.0.8
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.9
jupyterlab_server==2.25.0
keyring==23.5.0
kiwisolver==1.4.5
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lazy_loader==0.3
Levenshtein==0.23.0
loguru==0.7.0
lxml==4.9.3
markdown-it-py==3.0.0
markovify==0.9.4
MarkupSafe==2.1.3
matplotlib==3.8.2
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==3.0.2
more-itertools==10.1.0
mpmath==1.3.0
msgpack==1.0.7
msgpack-numpy-opentensor==0.5.0
multidict==6.0.4
multiprocess==0.70.15
munch==2.5.0
mypy-extensions==1.0.0
nbclassic==1.0.0
nbclient==0.9.0
nbconvert==7.11.0
nbformat==5.9.2
nest-asyncio==1.5.8
netaddr==0.10.1
nodeenv==1.8.0
notebook==6.5.5
notebook_shim==0.2.3
numpy==1.26.3
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.0
omegaconf==2.3.0
opencv-python==4.9.0.80
overrides==7.4.0
packaging==23.2
pandas==2.1.4
pandocfilters==1.5.0
parso==0.8.3
password-strength==0.0.3.post2
pathspec==0.12.1
pexpect==4.8.0
Pillow==9.3.0
platformdirs==4.1.0
pluggy==1.3.0
pre-commit==3.6.0
prometheus-client==0.18.0
prompt-toolkit==3.0.39
psutil==5.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
py-bip39-bindings==0.1.11
py-ed25519-zebra-bindings==1.0.1
py-sr25519-bindings==0.2.0
pyarrow==14.0.2
pyarrow-hotfix==0.6
pycparser==2.21
pycryptodome==3.19.1
pydantic==1.10.13
pydantic_core==2.14.6
Pygments==2.17.2
PyJWT==2.3.0
PyNaCl==1.5.0
pytest==7.4.4
pytest-asyncio==0.23.3
python-dateutil==2.8.2
python-dotenv==1.0.0
python-json-logger==2.0.7
python-Levenshtein==0.23.0
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==24.0.1
rapidfuzz==3.6.1
referencing==0.30.2
regex==2023.12.25
requests==2.31.0
resolvelib==0.8.1
retry==0.9.2
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.0
rpds-py==0.12.0
safetensors==0.4.1
scalecodec==1.2.7
scikit-image==0.22.0
scikit-learn==1.4.0
scipy==1.11.4
SecretStorage==3.3.1
Send2Trash==1.8.2
shtab==1.6.5
six==1.16.0
slowapi==0.1.9
sniffio==1.3.0
soupsieve==2.5
stack-data==0.6.3
starlette==0.27.0
substrate-interface==1.5.2
sympy==1.12
termcolor==2.4.0
terminado==0.17.1
tifffile==2023.12.9
tinycss2==1.2.1
tokenizers==0.15.0
tomli==2.0.1
toolz==0.12.0
torch==2.1.0
torchaudio==2.1.0
torchsde==0.2.6
torchvision==0.16.0
tornado==6.3.3
tqdm==4.66.1
traitlets==5.13.0
trampoline==0.1.2
transformers==4.36.2
triton==2.1.0
typer==0.9.0
types-python-dateutil==2.8.19.14
typing_extensions==4.9.0
tzdata==2023.4
Unidecode==1.3.8
uri-template==1.3.0
urllib3==2.1.0
uvicorn==0.22.0
virtualenv==20.25.0
wadllib==1.3.6
wcwidth==0.2.13
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
widgetsnbextension==4.0.9
xgboost==2.0.3
xxhash==3.4.1
yarl==1.9.4
yq==3.2.3
zipp==1.0.0
